+++ Documents/任务2-Tree类.txt
@@ -1,0 +1,11 @@
+任务二：文件夹哈希的实现
+包括Tree类
+Tree类
+数据域：
+1、	key：传入目录的key
+2、	value：传入目录的value
+
+方法
+1、	Tree(String path)：构造方法，传入路径，建立传入目录的key-value文件
+2、	setTreeKeyValue(String path)：深度优先遍历，为每个文件和文件夹建立key-value文件，并返回整个文件目录的key，用于构造方法中key的赋值
+3、	getTreeFileValue(String key)：传入一个目录的哈希key，返回其key-value文件中的内容，用于构造方法中value的赋值
